requirements.txt
onvif_zeep
requests
